examples/traces/requirements.txt
azure-ai-projects==1.0.0
azure-monitor-opentelemetry==1.6.13
opentelemetry-instrumentation-openai-v2==2.1b0